+++ src/chunking.hxx
@@ -10,8 +10,8 @@ namespace pcat
 	using substrate::fd_t;
 	using substrate::operator ""_KiB;
 
-	constexpr static size_t pageSize = 4_KiB;
-	constexpr static size_t transferBlockSize = 64 * pageSize;
+	constexpr static off_t pageSize = 4_KiB;
+	constexpr static off_t transferBlockSize = 64 * pageSize;
 	extern std::vector<fd_t> inputFiles;
 	extern fd_t outputFile;
